requirements.txt
requests==2.31.0
PyGithub==1.59.0
python-dotenv==1.0.0
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.1
networkx==3.1
